Remove `generateFromMetadata` and `getMetadata` usage.
